Enemy can now damage player on collision + removed insta-kill key
SInce the player has a way to die now, I removed the code where K was set to kill the player instantly

+++ TPP2/Player.h
@@ -43,7 +43,7 @@ class Player: public GameObject{
         void SetHealth(double h);
         void SetMaxHealth(double h);
 
-        void ChangeHealth(int i);
+        void ChangeHealth(double i);
 
         //overrides
         void Update() override;
